requirements.txt
Pillow @ file:///C:/ci/pillow_1625663286921/work
torch==1.9.0
torchvision==0.11.2
scikit-image==0.18.1
scikit-learn==1.0.1
natsort==8.0.0
